% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localboot.R
\name{local_boot}
\alias{local_boot}
\title{Local Bootstrap Function}
\usage{
local_boot(
  A,
  quantile_n = 0,
  B,
  returns = "boot",
  method = "own",
  distance = "zhu",
  kowning_u = NULL,
  induced_sampling = TRUE,
  weighted = FALSE,
  getT = NULL,
  user_blist = NULL,
  fast = NULL,
  ...
)
}
\arguments{
\item{A}{A matrix input.}

\item{quantile_n}{Quantile parameter, default is 0.}

\item{B}{Integer parameter.}

\item{returns}{Type of return value, default is "boot".}

\item{method}{Method to be used, default is "own".}

\item{distance}{Distance measurement, default is "zhu".}

\item{kowning_u}{Optional parameter for known user.}

\item{induced_sampling}{Boolean for induced sampling, default is TRUE.}

\item{weighted}{Boolean for weighted graph, default is FALSE.}

\item{getT}{Function to get T, default is NULL.}

\item{user_blist}{Optional user bootstrap list, default is NULL.}

\item{fast}{Boolean for fast computation, default is NULL.}

\item{...}{Additional arguments.}
}
\value{
Depends on the 'returns' parameter.
}
\description{
Local Bootstrap Function
}
\examples{
# Example usage
result <- local_boot(A = matrix_data, B = 100, returns = "boot")
}
