\name{localboot-package}
\alias{localboot-package}
\alias{localboot}
\docType{package}
\title{
\packageTitle{localboot}
}
\description{
\packageDescription{localboot}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{localboot}
\packageIndices{localboot}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{localboot}

Maintainer: \packageMaintainer{localboot}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
