% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_adj}
\alias{plot_adj}
\title{Plot Adjacency Matrix}
\usage{
plot_adj(X, ...)
}
\arguments{
\item{X}{An adjacency matrix to be plotted.}

\item{...}{Additional graphical parameters to pass to `image`.}
}
\value{
Generates a plot.
}
\description{
This function creates a plot of an adjacency matrix, where the matrix is displayed as an image.
}
\examples{
adj_matrix <- matrix(rbinom(100, 1, 0.5), 10, 10)
plot_adj(adj_matrix)

}
