% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_P}
\alias{plot_P}
\title{Plot Probability Matrix}
\usage{
plot_P(P)
}
\arguments{
\item{P}{A probability matrix to be plotted.}
}
\value{
Generates a filled contour plot.
}
\description{
This function creates a filled contour plot of a probability matrix using the `viridis` color palette.
The plot is created using the `ggplot2` and `viridis` libraries.
}
\examples{
P_matrix <- matrix(runif(100), 10, 10)
plot_P(P_matrix)

}
