% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localboot-package.R
\docType{package}
\name{localboot-package}
\alias{localboot-package}
\alias{_PACKAGE}
\title{localboot: A Package for Local Bootstrap Methods for Various Networks}
\description{
This package provides tools for bootstrapping and visualizing network data. 
It provide functions to bootstrap various networks via local bootstrap procedure.
Additionally, it includes functions for generating graphon probability matrices, creating 
network adjacency matrices from probability matrices, and plotting network 
structures. This package is a great tool to estiamte network uncertainty.
}
\author{
\strong{Maintainer}: Tianhai Zu \email{zuti@mail.uc.edu}

Other contributors:
\itemize{
  \item Yichen Qin \email{qinyn@ucmail.uc.edu} [contributor]
}

}
