% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localboot-package.R
\docType{package}
\name{localboot-package}
\alias{localboot-package}
\alias{_PACKAGE}
\title{localboot: A Package for Local Bootstrap Methods for Various Networks}
\description{
Network analysis usually requires estimating the uncertainty of
a graph statistics. Through this package, we provides tools to bootstrap 
various networks via local bootstrap procedure. Additionally, it includes 
functions for generating probability matrices, creating network adjacency 
matrices from probability matrices, and plotting network structures.
}
\author{
\strong{Maintainer}: Tianhai Zu \email{zuti@mail.uc.edu}

Authors:
\itemize{
  \item Yichen Qin \email{qinyn@ucmail.uc.edu} [contributor]
}

}
